fix test case

+++ Integral.Games.Tests/Formulae/TestLevelFormula.cs
@@ -2,10 +2,10 @@
 
 namespace Integral.Formulae
 {
-    internal sealed class TestLevelFormula : DeltaFormula<int>
+    internal sealed class TestLevelFormula : DeltaFormula<float>
     {
         private const float ExperiencePower = 2, ExperienceMultiplier = 100;
 
-        public int Evaluate(int previous, int current) => 1 + (int)Math.Truncate(Math.Pow(current / ExperienceMultiplier, 1 / ExperiencePower));
+        public float Evaluate(float previous, float current) => 1 + (float)Math.Truncate(Math.Pow(current / ExperienceMultiplier, 1 / ExperiencePower));
     }
 }
